Need to find a way to call the filter action after the setDate action is called.
