+++ front-end/src/actions/timelineActions.js
@@ -1,13 +1,3 @@
-import { getData } from './visualizationActions';
-
-export const setSelectedTime = selectedDates => (dispatch) => {
-  const postBody = {
-    ...selectedDates,
-  };
-
-  dispatch(getData(postBody));
-};
-
 export const setSelectedTimeline = () => (dispatch) => {
   const fetchData = {
     method: 'POST',
@@ -28,4 +18,8 @@ export const setSelectedTimeline = () => (dispatch) => {
     .catch((err) => {
       console.error.bind(err);
     });
+};
+
+export const setSelectedDate = selectedDates => (dispatch) => {
+  dispatch({ type: 'SET_DATE_RANGE', REPT_DT: { start: selectedDates.startDate, end: selectedDates.endDate } });
 };
