+++ front-end/src/reducers/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import demographic from './demographicReducer';
-import data from './visualizationReducer';
+import filters from './filterReducer';
+import mainVisualization from './visualizationReducer';
 import timeline from './timelineReducer';
 
-export default combineReducers({ demographic, data, timeline });
+export default combineReducers({ demographic, filters, mainVisualization, timeline });
