+++ front-end/src/reducers/visualizationReducer.js
@@ -1,11 +1,11 @@
 const initialVisualizationState = {
-  things: '',
+  data: '',
 };
 
 export default (state = initialVisualizationState, action = {}) => {
   switch (action.type) {
     case 'UPDATE_DATA':
-      return Object.assign({}, state, { things: action.things });
+      return Object.assign({}, state, { data: action.data });
     default: return state;
   }
 };
